LPS-66605 Move to internal package
