+++ modules/apps/knowledge-base/knowledge-base-item-selector-web/src/main/java/com/liferay/knowledge/base/item/selector/web/internal/KBAttachmentItemSelectorView.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.knowledge.base.item.selector.web;
+package com.liferay.knowledge.base.item.selector.web.internal;
 
 import com.liferay.item.selector.ItemSelectorReturnType;
 import com.liferay.item.selector.ItemSelectorView;
@@ -20,7 +20,7 @@ import com.liferay.item.selector.criteria.FileEntryItemSelectorReturnType;
 import com.liferay.item.selector.criteria.URLItemSelectorReturnType;
 import com.liferay.item.selector.criteria.UploadableFileReturnType;
 import com.liferay.knowledge.base.item.selector.criterion.KBAttachmentItemSelectorCriterion;
-import com.liferay.knowledge.base.item.selector.web.display.context.KBAttachmentItemSelectorViewDisplayContext;
+import com.liferay.knowledge.base.item.selector.web.internal.display.context.KBAttachmentItemSelectorViewDisplayContext;
 import com.liferay.portal.kernel.theme.ThemeDisplay;
 import com.liferay.portal.kernel.util.AggregateResourceBundleLoader;
 import com.liferay.portal.kernel.util.ListUtil;
